Do not remove trailing slashes of all URLs
According to RFC2518 5.2 a collection resource is identified by a trailing
slash. Although servers should treat the URI of a collection with and without a
trailing slash the same, some servers, e.g. Radicale, will not behave in that
way. Therefore trailing slashes should always be kept to be RFC compliant.